VMS-18586: Retry Mediaserver update installation if dpkg is locked by another process
Note: For installations with root tool enabled this won't work until 4.1 is installed to system,
because root tool did not expose dpkg exit status to shell. Now it does.

+++ vms/libs/nx_system_commands/src/nx/system_commands_linux.cpp
@@ -609,7 +609,7 @@ std::string SystemCommands::serializedDmiInfo()
         "/usr/sbin/dmidecode -t17",
         [&values, &prefixes, trim](const char* line)
         {
-            for (int index = 0; index < prefixes.size(); index++)
+            for (size_t index = 0; index < prefixes.size(); index++)
             {
                 const char* ptr = strstr(line, prefixes[index]);
                 if (ptr)
